@@ -3,11 +3,11 @@ using System;
 
 namespace Pscx.Properties {
     public static class PscxAssemblyInfo {
-        public const string AssemblyVersion = "3.6.4";
-        public const string FileVersion = "3.6.4";
+        public const string AssemblyVersion = "3.6.5";
+        public const string FileVersion = "3.6.5";
 
         public const string Product = "PowerShell Core Community Extensions";
-        public const string Copyright = "Copyright \xa9 2005-2023 Keith Hill, Oisin Grehan, Dan Luca, and contributors";
+        public const string Copyright = "Copyright \xa9 2005-2024 Keith Hill, Oisin Grehan, Dan Luca, and contributors";
         public const string Company = "https://github.com/pscx/Pscx; fork https://github.com/danluca/Pscx";
         public const string Description = "Extensions to PowerShell Core created by community - utilities, enhanced functionality, etc.";
     }
