@@ -31,5 +31,5 @@ using System.Runtime.InteropServices;
 //
 // You can specify all the values or you can default the Revision and Build Numbers 
 // by using the '*' as shown below:
-[assembly: AssemblyVersion("3.6.4")]
-[assembly: AssemblyFileVersion("3.6.4")]
+[assembly: AssemblyVersion("3.6.5")]
+[assembly: AssemblyFileVersion("3.6.5")]
