@@ -5,7 +5,7 @@ namespace PscxUnitTests.SimpleUnits {
     [TestFixture]
     public class LengthTests {
         private static void TestLength(string expected, double meters) {
-            Assert.AreEqual(expected, new Length(meters).ToString());
+            Assert.That(expected, Is.EqualTo(new Length(meters).ToString()));
         }
 
         [Test]
@@ -21,7 +21,7 @@ namespace PscxUnitTests.SimpleUnits {
         public void TestToAutoscaledString() {
             //for length, the autoscaled unit is matched to inch
             TestLength("3.937 in", 0.1);
-            Assert.AreEqual("10.000 cm", Length.FromMillimeters(100).ToUnit(Length.Centimeter).ToString());
+            Assert.That("10.000 cm", Is.EqualTo(Length.FromMillimeters(100).ToUnit(Length.Centimeter).ToString()));
             TestLength("1.000 km", 1000);
             TestLength("15.240 km", 15240);
         }
@@ -33,20 +33,20 @@ namespace PscxUnitTests.SimpleUnits {
             Length l3 = new(146.998);
             Length l4 = (Length)"2458.6 m";
 
-            Assert.AreEqual("7.837 km", l1.ToUnit(Length.Kilometer).AsString());
-            Assert.AreEqual("74.066 m", l2.ToString());
-            Assert.AreEqual("0.091 mi", l3.ToString(Length.Mile));
-            Assert.AreEqual("2.459 km", l4.ToString());
+            Assert.That("7.837 km", Is.EqualTo(l1.ToUnit(Length.Kilometer).AsString()));
+            Assert.That("74.066 m", Is.EqualTo(l2.ToString()));
+            Assert.That("0.091 mi", Is.EqualTo(l3.ToString(Length.Mile)));
+            Assert.That("2.459 km", Is.EqualTo(l4.ToString()));
         }
 
         [Test]
         public void TestConversion() {
-            Assert.AreEqual(Length.Meter, (Unit)"m");
-            Assert.AreEqual(Length.Millimeter, (Unit)"mm");
-            Assert.AreEqual(Length.Mile, (Unit)"mi");
+            Assert.That(Length.Meter, Is.EqualTo((Unit)"m"));
+            Assert.That(Length.Millimeter, Is.EqualTo((Unit)"mm"));
+            Assert.That(Length.Mile, Is.EqualTo((Unit)"mi"));
             var area = new Length(1.024, Length.Meter) * new Length(1.024, Length.Meter);
-            Assert.IsAssignableFrom(typeof(Area), area);
-            Assert.AreEqual(new Area(1.049, Area.SquareMeter).ToString(), area.ToString());
+            Assert.That(area, Is.AssignableTo(typeof(Area)));
+            Assert.That(new Area(1.049, Area.SquareMeter).ToString(), Is.EqualTo(area.ToString()));
         }
     }
 }
